Fashion_CV/create_dataset/pyproject.toml
[project]
name = "fashion-cv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.2.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "tqdm>=4.67.3",
]
